# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: trendsAF API Documentation
  description: |
    This API provides CRUD operations for trendsAF, allowing users to interact with the endpoints provided. Each endpoint provides typical HTTP methods such as GET, POST, PUT, and DELETE.
  version: 1.0.0
servers:
  - url: http://165.232.43.233:5080
paths:
...
  /admin/crops:
    post:
      summary: Add Crop
      description: Adds a new crop to the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                crop_category_id:
                  type: integer
                crop_name:
                  type: string
      responses:
        '200':
          description: Crop added successfully
  /admin/countries:
    post:
      summary: Add Country
      description: Adds a new country to the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                country_name:
                  type: string
                country_code:
                  type: string
      responses:
        '200':
          description: Country added successfully
  /admin/countries/regions:
    post:
      summary: Add Region
      description: Adds a new region to the specified country.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                country_id:
                  type: integer
                country_name:
                  type: string
                region_name:
                  type: string
      responses:
        '200':
          description: Region added successfully
...
  /admin/crops/process_state:
    post:
      summary: Add Crop Process State
      description: Adds a new process state for a specific crop.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                crop_id:
                  type: integer
                crop_variety_id:
                  type: integer
                process_state:
                  type: string
      responses:
        '200':
          description: Process state added successfully
  /admin/crops/variety:
    post:
      summary: Add Crop Variety
      description: Adds a new variety to a specific crop.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                crop_id:
                  type: integer
                crop_variety_name:
                  type: string
      responses:
        '200':
          description: Crop variety added successfully
  /admin/products:
    post:
      summary: Add Product
      description: Adds a new product remeber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                crop_id:
                  type: integer
                crop_variety_id:
                  type: integer
                country_id:
                  type: integer
                region_id:
                  type: integer
                price:
                  type: number
      responses:
        '200':
          description: New Product Price added successfully
  /admin/products/import:
    post:
      summary: Products Import
      description: Imports Product as a CSV file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file_id:
                  type: string

      responses:
        '200':
          description: Product imported successfully
  /admin/import:
    post:
      summary: Import products Product
      description: Imports a csv file for import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file_id:
                  type: string
      responses:
        '200':
          description: New Product Price added successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []